Extract CommodityLocationReader into its own class
